Minor refactoring regarding AbstractManufacturerHandler
